Teachers Controller
Create file and endpoint teacher.

+++ prisma/migrations/20211213022903_teachers/migration.sql
@@ -3,7 +3,7 @@ CREATE TABLE "Teacher" (
     "id" TEXT NOT NULL,
     "name" TEXT NOT NULL,
     "email" TEXT NOT NULL,
-    "cpf" BIGINT NOT NULL,
+    "cpf" INTEGER NOT NULL,
     "formation" TEXT NOT NULL,
 
     CONSTRAINT "Teacher_pkey" PRIMARY KEY ("id")
